Refaktorizace třídy Program a unit testů

+++ NNPTPZ1/ComplexNumber.cs
@@ -5,7 +5,7 @@ namespace NNPTPZ1.Mathematics
     public class ComplexNumber
     {
         public double RealPart { get; set; }
-        public float ImaginaryPart { get; set; }
+        public double ImaginaryPart { get; set; }
 
         public readonly static ComplexNumber Zero = new ComplexNumber()
         {
